fixed exit to return functions

+++ zsh/functions/brew.zsh
@@ -9,13 +9,23 @@ brew-restore() {
 }
 
 brew-install() {
-  : "${1? brew-install [options] formula|cask [...]}"
+  if [ -z "$1" ]; then
+      echo "Usage: brew-install [options] <formula|cask> [...]"
+      echo ""
+      return 1
+  fi
+
   brew install $1
   brew-backup
 }
 
 brew-uninstall() {
-  : "${1? brew-uninstall [options] installed_formula|installed_cask [...]}"
+  if [ -z "$1" ]; then
+      echo "Usage: brew-uninstall [options] <installed_formula|installed_cask> [...]}"
+      echo ""
+      return 1
+  fi
+
   brew uninstall $1
   brew-backup
 }
